# app/projects/phoenixA/openapi.yaml
openapi: 3.0.3
info:
  title: PhoenixA Data and Feature Platform API
  version: v1.40.0
  description: |
    PhoenixA data middle-platform API for AmazingData datasets and governed,
    versioned Feature Platform materializations.

    ## Design principle

    This OpenAPI spec describes the **API shape** (paths, parameters, response
    envelopes, error codes). It does NOT describe individual field meanings —
    those live in the field dictionary and are exposed via the discovery API.

    ## Two-contract model

    - **API contract** (this file): how to call — paths, params, envelope.
    - **Data contract** (discovery API): what fields exist, their meanings,
      units, types. Always query discovery first when you need field info.

    ## Agent integration flow

    1. `GET /api/v2/catalog/datasets` — list available datasets.
    2. `GET /api/v2/catalog/datasets/{dataset}/fields?type=X&search=keyword`
       — discover field names by keyword; learn raw_field, canonical_field,
       unit, scale, value_type, storage_location.
    3. `GET /api/v2/catalog/enums/{enum_name}` — resolve enum codes.
    4. `GET /api/v2/{dataset-query-path}?fields=RAW_A,RAW_B&format=flat` —
       query data, projecting only the fields you discovered.
    5. `POST /api/v2/catalog/field-coverage/scan` periodically — find SDK-added
       fields the dictionary hasn't caught up with.

    Unknown fields in `fields=` return 400 with suggestions from the
...
servers:
  - url: http://localhost:8085
...

...
paths:
...
  /api/v2/catalog/datasets:
    get:
      tags: [Discovery]
      summary: List governed datasets
      description: |
        Returns all datasets registered in the field dictionary, with their
        storage table, data_types, and query/discovery API path templates.
      parameters:
        - in: query
          name: source
          schema: { type: string }
          description: Filter by source (e.g. "amazing_data").
      responses:
        '200':
          description: Dataset list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetListResponse' }
  /api/v2/catalog/datasets/{dataset}/fields:
    get:
      tags: [Discovery]
      summary: Discover fields in a dataset
      description: |
        Returns field dictionary rows for one dataset, optionally filtered by
        data_type, search keyword, and metadata flags. Each row carries the
        raw_field name (what to pass to `fields=` in queries), canonical_field,
        label_zh, unit, scale, value_type, storage_location, is_core, and
        enum_ref.
      parameters:
        - in: path
          name: dataset
          required: true
          schema: { type: string }
          description: Dataset name (e.g. "financial_statement").
        - in: query
          name: source
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
          description: Data type filter (e.g. "balance_sheet", "dividend").
        - in: query
          name: search
          schema: { type: string }
          description: Substring match on raw_field / canonical_field / label_zh / aliases.
        - in: query
          name: include
          schema: { type: string, enum: [all, core, metadata] }
          description: Filter by importance. Default "all".
      responses:
        '200':
          description: Field list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldDiscoveryResponse' }
  /api/v2/catalog/enums/{enum_name}:
    get:
      tags: [Discovery]
      summary: Resolve enum values
      description: Returns all code/label pairs for a named enum (e.g. REPORT_TYPE).
      parameters:
        - in: path
          name: enum_name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Enum values.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EnumDiscoveryResponse' }
        '404':
          description: Enum not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }

  # ─── Discovery: per-security data coverage ───
  /api/v2/catalog/securities/{security_id}/datasets/summary:
    get:
      tags: [Discovery]
      summary: Per-security data coverage summary
      description: |
        Returns row counts and time ranges for each dataset/data_type that
        exists for a given security_id. Covers financial_statement (by
        statement_type × report_type), corporate_action (by action_type),
        and equity_structure. Generic discovery API — not BI-specific.
      parameters:
        - in: path
          name: security_id
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
          description: security_registry.id (Phase 3 surrogate key).
      responses:
        '200':
          description: Coverage summary.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SecurityCoverageResponse' }

  # ─── Discovery: field coverage ───
...
  /api/v2/catalog/field-coverage:
    get:
      tags: [Coverage]
      summary: List field coverage observations
      parameters:
        - in: query
          name: dataset
          schema: { type: string }
        - in: query
          name: source
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [governed, ungoverned] }
          description: Filter by status. Omit for all.
      responses:
        '200':
          description: Observations.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CoverageListResponse' }

  # ─── Query: financial statements ───
  /api/v2/financial/{source}/{statement_type}:
    get:
      tags: [Query]
      summary: Query financial statements
      description: |
        Dictionary-resolved query against the financial_statement table.
        Fields in `fields=` are resolved through the field dictionary — pass
        raw_field names (e.g. TOTAL_ASSETS) or canonical_field names; mixing
        is allowed. Unknown fields return 400 with suggestions.

        Use discovery (`/api/v2/catalog/datasets/financial_statement/fields`)
        to learn available field names, units, and types.
      parameters:
        - in: path
          name: source
          required: true
          schema: { type: string }
          example: amazing_data
        - in: path
          name: statement_type
          required: true
          schema: { type: string, enum: [balance_sheet, income, cashflow, profit_express, profit_notice] }
        - in: query
          name: format
          schema: { type: string, enum: [flat, nested], default: nested }
          description: |
            - `nested` (default): top_level columns + full data_json object.
            - `flat`: returns projected fields as top-level keys.
        - in: query
          name: fields
          schema: { type: string }
          description: |
            Comma-separated raw/canonical field names for flat projection.
            Ignored for nested format (which returns full data_json).
            Unknown fields return 400 with suggestions.
        - in: query
          name: security_id
          schema: { type: integer, format: int64, minimum: 1 }
          description: Single security_id (Phase 3 surrogate key) to filter on.
        - in: query
          name: security_ids
          schema: { type: string }
          description: Comma-separated list of security_ids.
        - in: query
          name: report_period
          schema: { type: string }
          description: Exact match on report_period (YYYY-MM-DD).
        - in: query
          name: period_start
          schema: { type: string }
        - in: query
          name: period_end
          schema: { type: string }
        - in: query
          name: ann_date_before
          schema: { type: string }
          description: PIT filter — only rows announced before this date.
        - in: query
          name: statement_code
          schema: { type: string }
          description: Numeric code (1=合并, 2=合并单季, 6=母公司, 7=母公司单季).
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
      responses:
        '200':
          description: Query result.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/FlatQueryResponse'
                  - $ref: '#/components/schemas/NestedQueryResponse'
        '400':
          description: Unknown field in `fields=`. Body contains suggestions.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldResolutionError' }
  /api/v2/financial/{source}/{statement_type}/upsert:
    post:
      tags: [Write]
      summary: Batch upsert financial statements
      parameters:
        - in: path
          name: source
          required: true
          schema: { type: string }
        - in: path
          name: statement_type
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: array, items: { type: object } }
      responses:
        '200':
          description: Upserted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UpsertResponse' }

  # ─── Query: corporate actions ───
  /api/v2/corporate-action/{source}/{action_type}:
    get:
      tags: [Query]
      summary: Query corporate actions (dividend / right_issue)
      parameters:
        - in: path
          name: source
          required: true
          schema: { type: string }
        - in: path
          name: action_type
          required: true
          schema: { type: string, enum: [dividend, right_issue] }
        - in: query
          name: format
          schema: { type: string, enum: [flat, nested], default: nested }
        - in: query
          name: fields
          schema: { type: string }
          description: Comma-separated raw/canonical field names for flat projection.
        - in: query
          name: security_id
          schema: { type: integer, format: int64, minimum: 1 }
          description: Single security_id (Phase 3 surrogate key) to filter on.
        - in: query
          name: security_ids
          schema: { type: string }
          description: Comma-separated list of security_ids.
        - in: query
          name: report_period
          schema: { type: string }
        - in: query
          name: period_start
          schema: { type: string }
        - in: query
          name: period_end
          schema: { type: string }
        - in: query
          name: ann_date_before
          schema: { type: string }
        - in: query
          name: progress_code
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
      responses:
        '200':
          description: Query result.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/FlatQueryResponse'
                  - $ref: '#/components/schemas/NestedQueryResponse'
        '400':
          description: Unknown field in `fields=`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldResolutionError' }
  /api/v2/corporate-action/{source}/{action_type}/upsert:
    post:
      tags: [Write]
      summary: Batch upsert corporate actions
      parameters:
        - in: path
          name: source
          required: true
          schema: { type: string }
        - in: path
          name: action_type
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: array, items: { type: object } }
      responses:
        '200':
          description: Upserted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UpsertResponse' }

  # ─── Query: equity structure ───
  /api/v2/equity-structure/{source}:
    get:
      tags: [Query]
      summary: Query equity structure
      description: |
        AmazingData get_equity_structure output. Has a single data_type
        (equity_structure) — no `type` parameter needed.
      parameters:
        - in: path
          name: source
          required: true
          schema: { type: string }
        - in: query
          name: format
          schema: { type: string, enum: [flat, nested], default: nested }
        - in: query
          name: fields
          schema: { type: string }
        - in: query
          name: security_id
          schema: { type: integer, format: int64, minimum: 1 }
          description: Single security_id (Phase 3 surrogate key) to filter on.
        - in: query
          name: security_ids
          schema: { type: string }
          description: Comma-separated list of security_ids.
        - in: query
          name: change_date
          schema: { type: string }
          description: Exact match.
        - in: query
          name: change_start
          schema: { type: string }
        - in: query
          name: change_end
          schema: { type: string }
        - in: query
          name: ann_date_before
          schema: { type: string }
        - in: query
          name: current_only
          schema: { type: string, enum: ["1", "0", "true", "false"] }
          description: Only rows with current_sign=1.
        - in: query
          name: valid_only
          schema: { type: string, enum: ["1", "0", "true", "false"] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
      responses:
        '200':
          description: Query result.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/FlatQueryResponse'
                  - $ref: '#/components/schemas/NestedQueryResponse'
        '400':
          description: Unknown field in `fields=`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FieldResolutionError' }
  /api/v2/equity-structure/{source}/upsert:
    post:
      tags: [Write]
      summary: Batch upsert equity structure rows
      parameters:
        - in: path
          name: source
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: array, items: { type: object } }
      responses:
        '200':
          description: Upserted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UpsertResponse' }

  # --- Feature Platform: registry and governance ---
  /api/v2/features/registry/sync:
    post:
      tags: [Feature Platform]
      summary: Synchronize governed feature manifests
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeatureRegistrySyncRequest' }
      responses:
        '200':
          description: Per-manifest synchronization result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeatureRegistrySyncResponse' }
        '400': { $ref: '#/components/responses/FeatureError' }
        '409': { $ref: '#/components/responses/FeatureError' }
        '422': { $ref: '#/components/responses/FeatureError' }
  /api/v2/features/definitions:
    get:
      tags: [Feature Platform]
      summary: List governed feature definitions
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: owner, schema: { type: string } }
        - $ref: '#/components/parameters/FeatureLimit'
        - $ref: '#/components/parameters/FeatureOffset'
      responses:
        '200':
          description: Feature definition page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeaturePage' }
  /api/v2/features/definitions/{feature_code}:
    parameters:
      - $ref: '#/components/parameters/FeatureCode'
    get:
      tags: [Feature Platform]
      summary: Get a definition and all version summaries
      responses:
        '200':
          description: Definition detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeatureDefinitionDetail' }
        '404': { $ref: '#/components/responses/FeatureError' }
  /api/v2/features/definitions/{feature_code}/lifecycle-events:
    parameters:
      - $ref: '#/components/parameters/FeatureCode'
    get:
      tags: [Feature Platform]
      summary: List audited lifecycle events for a feature
      parameters:
        - $ref: '#/components/parameters/FeatureLimit'
      responses:
        '200':
          description: Most recent lifecycle events.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeatureLifecycleEventPage' }
        '400': { $ref: '#/components/responses/FeatureError' }
  /api/v2/features/versions/{version_id}:
    parameters:
      - $ref: '#/components/parameters/FeatureVersionID'
    get:
      tags: [Feature Platform]
      summary: Get a feature version by registry ID
      responses:
        '200':
          description: Version, implementation, and dependency snapshot.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeatureVersionSummary' }
        '404': { $ref: '#/components/responses/FeatureError' }
  /api/v2/features/definitions/{feature_code}/versions/{version}:publish:
    parameters:
      - $ref: '#/components/parameters/FeatureCode'
      - $ref: '#/components/parameters/FeatureVersionNumber'
    post:
      tags: [Feature Platform]
      summary: Publish an immutable feature version
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeatureLifecycleTransitionRequest' }
      responses:
        '200':
          description: Version published.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeatureLifecycleResponse' }
        '404': { $ref: '#/components/responses/FeatureError' }
        '400': { $ref: '#/components/responses/FeatureError' }
        '409': { $ref: '#/components/responses/FeatureError' }
        '422': { $ref: '#/components/responses/FeatureError' }
  /api/v2/features/definitions/{feature_code}/versions/{version}:deprecate:
    parameters:
      - $ref: '#/components/parameters/FeatureCode'
      - $ref: '#/components/parameters/FeatureVersionNumber'
    post:
      tags: [Feature Platform]
      summary: Deprecate a published feature version
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeatureLifecycleTransitionRequest' }
      responses:
        '200':
          description: Version deprecated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeatureLifecycleResponse' }
        '404': { $ref: '#/components/responses/FeatureError' }
        '400': { $ref: '#/components/responses/FeatureError' }
        '409': { $ref: '#/components/responses/FeatureError' }
  /api/v2/features/lineage/{feature_code}:
    parameters:
      - $ref: '#/components/parameters/FeatureCode'
    get:
      tags: [Feature Platform]
      summary: Get versioned upstream and downstream lineage
      responses:
        '200':
          description: Version-level feature and data-field lineage.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeatureLineage' }
        '404': { $ref: '#/components/responses/FeatureError' }
...
  /api/v1/atlas-graph/search:
    get:
      tags: [Atlas Graph]
      summary: Search projected Atlas entities
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 1000 } }
      responses:
        '200':
          description: Projected entity list.
...
components:
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
...